fix: 메서드들에서 id를 Long으로 처리함

+++ src/com/programmers/demo/common/repository/IMovieManager.java
@@ -8,9 +8,9 @@ import java.util.List;
 public interface IMovieManager {
     void addMovie(Movie movie); // 영화 등록
 
-    void deleteMovie(int id); // 영화 삭제
+    void deleteMovie(Long id); // 영화 삭제
 
-    void updateMovie(int id, LocalDate startDate, LocalDate endDate); // 영화 수정(날짜만 가능)
+    void updateMovie(Long id, LocalDate startDate, LocalDate endDate); // 영화 수정(날짜만 가능)
 
     Movie getMovie(String title); //제목으로 영화 검색
